requirements.txt
alembic==1.0.7
amqp==2.4.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Migrate==2.4.0
flask-nameko==1.4.0
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
idna==2.8
importlib-metadata==0.8
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.3.0
Mako==1.0.7
MarkupSafe==1.1.1
mock==2.0.0
monotonic==1.5
nameko==2.11.0
nameko-sqlalchemy==1.5.0
path.py==11.5.0
pbr==5.1.3
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==3.13
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.18
sqlalchemy-migrate==0.12.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
wrapt==1.11.1
zipp==0.3.3
